enterprise ai suite/step1_secure_chat/requirements.txt
streamlit
sqlalchemy
passlib[bcrypt]
python-jose
bcrypt==4.0.1
